Consolidate browse controller/module
